Split rank system from object file
Not really necessary, but ideally I'd move that to the script part and
make it an OC specific thing.
